module5-task1 Добавляет разделение js кода на модули
