+++ js/main.js
@@ -1,122 +1,3 @@
-// Определение констант
-const constants = {
-  //Количество объектов изображений
-  PICTURES_COUNT: 25,
-  //Минимальное количество лайков
-  MINIMUM_COUNT_LIKES: 15,
-  //Максимальное количество лайков
-  MAXIMUM_COUNT_LIKES: 200
-};
+import {getPosts} from './data';
 
-//Массив описаний
-const DESCRIPTIONS = [
-  'Фотографии являются окном в прошлое, позволяя нам заглянуть в моменты, которые уже ушли',
-  'Они могут заставить нас увидеть привычные вещи с новой перспективы',
-  'Фотографии являются уникальным способом сохранить воспоминания',
-  'Через фотографии мы можем поделиться своими историями и переживаниями',
-  'Фотографии способны вызывать эмоции и могут быть сильным инструментом коммуникации',
-  'Они помогают нам увидеть красоту в простых вещах и моментах',
-  'Фотографии могут донести глубокий смысл и идеи без слов',
-  'Они позволяют нам остановить время и сохранить его в кадре',
-  'Фотографии могут служить документацией истории и культуры',
-  'Хорошая фотография требует терпения, творчества и внимания к деталям',
-  'Через фотографии мы можем увидеть красоту разных культур и традиций',
-  'Они могут быть способом выражения себя и своих чувств',
-  'Фотографии могут помочь нам увидеть мир через глаза других людей',
-  'Они могут стать мощным источником вдохновения и творчества',
-  'Фотографии могут запечатлеть и фиксировать изменения в мире и окружающей среде',
-  'Они позволяют нам по-новому взглянуть на себя и свою жизнь',
-  'Фотографии могут объединять людей и создавать связи через общие впечатления',
-  'Они могут рассказывать истории и вызывать диалог о важных темах',
-  'Фотографии могут вызывать у зрителя разнообразные чувства: радость, грусть, восторг, ностальгию и другие',
-  'Они могут отражать индивидуальность фотографа и его взгляд на мир',
-  'Фотографии могут стать наставником, вдохновляющим на собственное творчество',
-  'Они могут быть инструментом обращения внимания на социальные проблемы и вызов к действию',
-  'Фотографии могут открывать новые возможности для самовыражения и самореализации',
-  'Они являются важным источником информации для исследования и анализа различных явлений',
-  'Фотографии могут создавать мосты между людьми разных культур и национальностей',
-  'Они могут быть способом сохранить уникальные моменты и события в жизни',
-  'Фотографии могут отразить дух времени и атмосферу эпохи',
-  'Они могут подавать пример красоты, гармонии и совершенства в окружающем мире',
-  'Фотографии могут быть способом выразить свои чувства и эмоции через изображения',
-  'Они могут помогать нам замечать красоту и уникальность в повседневных моментах',
-  'Фотографии могут служить источником вдохновения для других видов искусства',
-  'Они могут вызывать у зрителя рефлексию и побуждать к изменениям',
-  'Фотографии могут помогать преодолевать границы языка и культуры, объединяя людей на уровне общего восприятия',
-  'Они могут быть олицетворением красоты, эмоций, истории и личных воспоминаний',
-  'Фотографии могут отразить сложность и многогранность мира, поднимая различные темы и вопросы',
-  'Они могут быть способом сохранить уникальные моменты жизни, которые быстро уходят в прошлое',
-  'Фотографии способны заставить задуматься и пересматривать свои стереотипы и представления о мире',
-  'Они могут передавать настроение и эмоциональное состояние автора и зрителя',
-  'Фотографии могут быть дорогой памятью о прошлом и источником вдохновения на будущее',
-  'Они могут стать способом зарядиться положительной энергией и радостью, перенося зрителя в удивительный мир изображений'
-];
-
-//Массив сообщений
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-//Массив имен
-const NAMES = [
-  'Александр', 'Михаил', 'Дмитрий', 'Иван', 'Сергей', 'Андрей', 'Николай', 'Егор', 'Павел',
-  'Алёна', 'Диана', 'Галина', 'Ксения', 'Елена', 'Ольга', 'Раиса', 'Анастасия', 'Вера', 'Татьяна',
-  'Арсений', 'Станислав', 'Леонид', 'Виктор', 'Фёдор', 'Пётр', 'Георгий', 'Борис', 'Владислав', 'Максим',
-  'Дарина', 'Регина', 'Оксана', 'Инга', 'Софья', 'Светлана', 'Юлия', 'Елена', 'Ангелина', 'Валя'
-];
-
-//Функция генерации случайного числа в заданном диапазоне
-const getRandomPositiveNumber = (firstNumber, secondNumber) => {
-  const lower = Math.ceil(Math.min(Math.abs(firstNumber), Math.abs(secondNumber)));
-  const upper = Math.floor(Math.max(Math.abs(firstNumber), Math.abs(secondNumber)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-//Получить случайное количество, случайных элементов из массива
-const getRandomElementByArray = (count, array) => Array.from({length: getRandomPositiveNumber(1, count)}, () => array[getRandomPositiveNumber(0, array.length - 1)]);
-
-//Функция генерации последовательных чисел
-const generateNumber = () => {
-  let id = 1;
-  return () => id++;
-};
-
-//Получить id объекта комментария
-const getCommentId = generateNumber();
-
-//Функция создания комментарий
-const createComments = () => ({
-  id: getCommentId(),
-  avatar: `img/avatar-${getRandomPositiveNumber(1, 6)}.svg`,
-  message: getRandomElementByArray(7, MESSAGES),
-  name: getRandomElementByArray(1, NAMES)
-});
-
-//Получить комментарии
-const getComments = (count) => Array.from({length: getRandomPositiveNumber(1, count)}, () => createComments());
-
-//Получить id для картинки
-const getPictureId = generateNumber();
-
-//Получить номер для картинки
-const getPhotoNumber = generateNumber();
-
-//Функция создания картинки
-const createPicture = () => ({
-  id: getPictureId(),
-  url: `photos/${getPhotoNumber()}.jpg`,
-  description: getRandomElementByArray(5, DESCRIPTIONS),
-  likes: getRandomPositiveNumber(constants.MINIMUM_COUNT_LIKES, constants.MAXIMUM_COUNT_LIKES),
-  comments: getComments(5)
-});
-
-//Получить объекты изображений
-const getPictures = () => Array.from({length: constants.PICTURES_COUNT}, () => createPicture());
-
-getPictures();
+getPosts();
